refactor: 💡 make navbar less complex, split up into components
